requirements.txt
neo4j>=5.14.0
neo4j-graphrag>=1.0.0
openai>=1.18.0
numpy>=1.24.0
mcp>=1.0.0
python-dotenv>=1.0.0 